requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
langdetect>=1.0.9
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
vaderSentiment>=3.3.2
snownlp>=0.12.3
pandas>=2.0.0
numpy>=1.24.0
